optimize getting block device + fix backup non-root mount points

+++ makebackup.sh
@@ -17,7 +17,11 @@ KEEPCOUNT=${3:-${KEEPCOUNT:-10}}
 
 TOP="$(btrfs subvolume show -b "$ROOT" | grep -Po '(?<=^\tTop level ID: \t\t)(\d+)$')"
 (($?)) && exit 1
-BLOCK="$(df "$ROOT" | head -n2 | tail -n1 | cut -f1 -d' ')"
+BLOCK="$(awk '$2=="'"$ROOT"'"{print$1}' /proc/mounts)"
+if (($?)) || [[ -z $BLOCK ]]; then
+    echo "cant find block device for $ROOT" >&2
+    exit 1
+fi
 
 ROOT_PATH="$(mktemp -d /tmp/makebackup.XXXXX)"
 trap "rmdir '$ROOT_PATH'" EXIT
@@ -39,8 +43,8 @@ else
     fi
 fi
 
-btrfs subvolume list --sort=-path "$ROOT" | awk '$9 ~ /^@'"$PREFIX"'-/ { print $2 " " $9 }' | awk "NR>$KEEPCOUNT" | while read -r LINE; do
+btrfs subvolume list --sort=-path "$ROOT" | awk '$9~/^@'"$PREFIX"'-/{print$2" "$9}' | awk "NR>$KEEPCOUNT" | while read -r LINE; do
     ID=$(echo "$LINE" | grep -Po '^\d+')
     NAME=$(echo "$LINE" | grep -Po '(?<= ).*')
-    btrfs subvolume delete -i "$ID" /
+    btrfs subvolume delete -i "$ID" "$ROOT"
 done
